Decouple discoverer from generator
