+++ cli/Cargo.toml
@@ -15,3 +15,5 @@ categories = ["database"]
 seaography-generator = { version = "^0.1", path = "../generator" }
 clap = { version = "3.2.20", features = ["derive"] }
 async-std = { version = "1.12.0", features = [ "attributes", "tokio1" ] }
+seaography-discoverer = { version = "^0.1", path = "../discoverer" }
+url = "2.2.2"
