+++ generator/Cargo.toml
@@ -13,10 +13,9 @@ categories = ["database"]
 
 [dependencies]
 quote = "1.0.21"
-url = "2.2.2"
 proc-macro2 = "1.0.43"
 syn = { version = "1.0.99", features = ["full"] }
 heck = "0.4.0"
 itertools = "0.10.3"
 sea-orm-codegen = { version = "0.9.2"}
-seaography-discoverer = { version = "^0.1", path = "../discoverer" }
+sea-query = { version = "0.26.3", default-features = false }
