+++ generator/src/sea_orm_codegen.rs
@@ -1,9 +1,9 @@
 use crate::util::add_line_break;
 
-pub type EntityHashMap = std::collections::HashMap<String, proc_macro2::TokenStream>;
+pub type EntityHashMap = std::collections::BTreeMap<String, proc_macro2::TokenStream>;
 
 pub fn generate_entities(
-    table_crate_stmts: Vec<seaography_discoverer::sea_schema::sea_query::TableCreateStatement>,
+    table_crate_stmts: Vec<sea_query::TableCreateStatement>,
     expanded_format: bool,
 ) -> crate::Result<EntityHashMap> {
     let entity_writer = sea_orm_codegen::EntityTransformer::transform(table_crate_stmts)?;
